+++ src/core_modules/capture-core/components/WidgetEnrollment/MapModal/MapModal.container.js
@@ -3,17 +3,17 @@ import React, { useCallback } from 'react';
 import { useGeometry } from '../hooks/useGeometry';
 import type { MapModalProps } from './MapModal.types';
 import { MapModal as MapModalComponent } from './MapModal.component';
-
-const DEFAULT_CENTER = [51.505, -0.09];
+import { useCenterPoint } from './hooks';
 
 export const MapModal = ({
     enrollment,
     onUpdate,
     setOpenMap,
     defaultValues,
-    center,
+    center: storedCenter,
 }: MapModalProps) => {
     const { geometryType, dataElementType } = useGeometry(enrollment);
+    const { center } = useCenterPoint(enrollment.orgUnit, storedCenter);
 
     const onSetCoordinates = useCallback((coordinates) => {
         const geometry = coordinates ? { type: geometryType, coordinates } : null;
@@ -22,7 +22,7 @@ export const MapModal = ({
 
     return (
         <MapModalComponent
-            center={center || DEFAULT_CENTER}
+            center={center}
             type={dataElementType}
             setOpen={setOpenMap}
             onSetCoordinates={onSetCoordinates}