+++ src/core_modules/capture-core/components/WidgetEventEdit/EditEventDataEntry/EditEventDataEntry.component.js
@@ -35,7 +35,6 @@ import {
     VirtualizedSelectField,
 } from '../../FormFields/New';
 import { statusTypes, translatedStatusTypes } from '../../../events/statusTypes';
-import { inMemoryFileStore } from '../../DataEntry/file/inMemoryFileStore';
 import labelTypeClasses from '../DataEntry/dataEntryFieldLabels.module.css';
 import { withDeleteButton } from '../DataEntry/withDeleteButton';
 import { withAskToCreateNew } from '../../DataEntry/withAskToCreateNew';
@@ -254,6 +253,7 @@ const buildGeometrySettingsFn = () => ({
                 label: i18n.t('Area'),
                 dialogLabel: i18n.t('Area'),
                 required: false,
+                orgUnit: props.orgUnit,
             });
         }
         return createComponentProps(props, {
@@ -261,6 +261,7 @@ const buildGeometrySettingsFn = () => ({
             label: i18n.t('Coordinate'),
             dialogLabel: i18n.t('Coordinate'),
             required: false,
+            orgUnit: props.orgUnit,
         });
     },
     getPropName: () => 'geometry',
@@ -356,7 +357,7 @@ const AOCSettings = {
 };
 
 const saveHandlerConfig = {
-    onIsCompleting: (props: Object) => props.completeDataEntryFieldValue,
+    onIsCompleting: (props: Object) => props.completeDataEntryFieldValue === 'true',
     onFilterProps: (props: Object) => {
         const { completeDataEntryFieldValue, ...passOnProps } = props;
         return passOnProps;
@@ -446,10 +447,6 @@ class EditEventDataEntryPlain extends Component<Props, State> {
         this.onHandleSwitchTab = this.onHandleSwitchTab.bind(this);
     }
 
-    componentWillUnmount() {
-        inMemoryFileStore.clear();
-    }
-
     onHandleSwitchTab = newMode => this.setState({ mode: newMode })
 
     renderScheduleView() {
@@ -524,6 +521,7 @@ class EditEventDataEntryPlain extends Component<Props, State> {
                 onSaveAndCompleteEnrollment={onSaveAndCompleteEnrollment(orgUnit)}
                 fieldOptions={this.fieldOptions}
                 dataEntrySections={this.dataEntrySections}
+                orgUnit={orgUnit}
                 programId={programId}
                 selectedOrgUnitId={orgUnit?.id}
                 {...passOnProps}