+++ src/core_modules/capture-core/components/WidgetProfile/DataEntry/DataEntry.container.js
@@ -107,6 +107,7 @@ export const DataEntry = ({
                 onGetValidationContext={getValidationContext}
                 errorsMessages={errorsMessages}
                 warningsMessages={warningsMessages}
+                orgUnit={{ id: orgUnitId }}
             />
         )
     );