+++ src/core_modules/capture-core/components/WidgetRelatedStages/FormComponents/index.js
@@ -1,0 +1,4 @@
+// @flow
+
+export { DateFieldForRelatedStages } from './DateFieldForRelatedStages';
+export { OrgUnitSelectorForRelatedStages } from './OrgUnitSelectorForRelatedStages';