+++ src/core_modules/capture-core/components/WidgetRelatedStages/RelatedStagesActions/index.js
@@ -1,0 +1,3 @@
+// @flow
+export { RelatedStagesActions } from './RelatedStagesActions.component';
+export type { ErrorMessagesForRelatedStages } from './RelatedStagesActions.types';