+++ src/epics/trackerCapture.epics.js
@@ -182,7 +182,7 @@ import {
 } from '../core_modules/capture-core/components/DataEntries/TeiRegistrationEntry/TeiRegistrationEntry.epics';
 import {
     completeSavingNewTrackedEntityInstanceEpic,
-    completeSavingNewTrackedEntityInstanceWithEnrollmentEpic,
+    completeSavingNewTrackedEntityInstanceWithEnrollmentEpic, failedSavingNewTrackedEntityInstanceWithEnrollmentEpic,
     startSavingNewTrackedEntityInstanceEpic,
     startSavingNewTrackedEntityInstanceWithEnrollmentEpic,
 } from '../core_modules/capture-core/components/Pages/New/RegistrationDataEntry/RegistrationDataEntry.epics';
@@ -208,8 +208,6 @@ import {
 import {
     saveNewEventSucceededEpic,
     saveNewEventFailedEpic,
-    saveEventAndCompleteEnrollmentSucceededEpic,
-    saveEventAndCompleteEnrollmentFailedEpic,
 } from '../core_modules/capture-core/components/Pages/EnrollmentAddEvent/EnrollmentAddEventPage.epics';
 import {
     updateEventSucceededEpic,
@@ -219,7 +217,6 @@ import {
     runRulesOnUpdateDataEntryFieldForNewEnrollmentEventEpic,
     runRulesOnUpdateFieldForNewEnrollmentEventEpic,
     saveNewEnrollmentEventEpic,
-    saveEventAndCompleteEnrollmentEpic,
     addNoteForNewEnrollmentEventEpic,
 } from '../core_modules/capture-core/components/WidgetEnrollmentEventNew';
 import {
@@ -341,6 +338,7 @@ export const epics = combineEpics(
     startSavingNewTrackedEntityInstanceWithEnrollmentEpic,
     completeSavingNewTrackedEntityInstanceEpic,
     completeSavingNewTrackedEntityInstanceWithEnrollmentEpic,
+    failedSavingNewTrackedEntityInstanceWithEnrollmentEpic,
     changedEnrollmentIdEpic,
     fetchEnrollmentIdEpic,
     verifyEnrollmentIdSuccessEpic,
@@ -365,9 +363,6 @@ export const epics = combineEpics(
     saveNewEnrollmentEventEpic,
     saveNewEventSucceededEpic,
     saveNewEventFailedEpic,
-    saveEventAndCompleteEnrollmentEpic,
-    saveEventAndCompleteEnrollmentSucceededEpic,
-    saveEventAndCompleteEnrollmentFailedEpic,
     updateEventSucceededEpic,
     updateEventFailedEpic,
     addNoteForNewEnrollmentEventEpic,